requirements.txt
requests==2.22.0
python-dateutil==2.8.1